+++ ObservatoryFramework/Files/Journal/Other/FighterRebuilt.cs
@@ -3,6 +3,6 @@
     public class FighterRebuilt : JournalBase
     {
         public string Loadout { get; init; }
-        public int ID { get; init; }
+        public ulong ID { get; init; }
     }
 }
