+++ ObservatoryFramework/Files/Journal/Other/NpcCrewRank.cs
@@ -4,7 +4,7 @@ namespace Observatory.Framework.Files.Journal
 {
     public class NpcCrewRank : JournalBase
     {
-        public int NpcCrewId { get; init; }
+        public ulong NpcCrewId { get; init; }
         public string NpcCrewName { get; init; }
         public RankCombat RankCombat { get; init; }
     }
