+++ ObservatoryFramework/Files/Journal/StationServices/FetchRemoteModule.cs
@@ -2,11 +2,11 @@
 {
     public class FetchRemoteModule : JournalBase
     {
-        public int ShipID { get; init; }
+        public ulong ShipID { get; init; }
         public int StorageSlot { get; init; }
         public string StoredItem { get; init; }
         public string StoredItem_Localised { get; init; }
-        public long ServerId { get; init; }
+        public ulong ServerId { get; init; }
         public long TransferCost { get; init; }
         public string Ship { get; init; }
         public long TransferTime { get; init; }
